pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zta-operator"
version = "0.1.0"
description = "ZeroTrustApplication Kubernetes Operator"
requires-python = ">=3.11"
dependencies = [
  "kopf==1.37.4",
  "kubernetes==30.1.0",
  "PyYAML==6.0.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
